fix the bugs and rename some vars
