--- TaskTimerApp/.idea/libraries/Gradle__androidx_customview_customview_1_0_0_aar.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Gradle: androidx.customview:customview:1.0.0@aar">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/007fb9606056faf85efea1381d5a3092/transformed/customview-1.0.0/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/007fb9606056faf85efea1381d5a3092/transformed/customview-1.0.0/res" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/007fb9606056faf85efea1381d5a3092/transformed/customview-1.0.0/AndroidManifest.xml" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/androidx.customview/customview/1.0.0/61f6a717d144dff3a6bda413d9abeeb2bca71581/customview-1.0.0-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
