+++ TaskTimerApp/app/src/main/java/com/example/tasktimerapp/fragments/HomeScreenFragment.kt
@@ -1,11 +1,11 @@
-package com.example.tasktimer.fragments
+package com.example.tasktimerapp.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.tasktimer.R
+import com.example.tasktimerapp.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
