+++ TaskTimerApp/app/src/main/res/values/strings.xml
@@ -1,5 +1,6 @@
 <resources>
     <string name="app_name">TaskTimer</string>
-    <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="preference_file_key">com.example.tasktimerapp.PREFERENCE_FILE_KEY</string>
+
 </resources>
